Move query definitions into their own namespace
